+++ jest-setup.ts
@@ -1,13 +1,9 @@
+/* eslint-disable no-console,jest/no-standalone-expect */
 import '@testing-library/jest-dom';
 import '@testing-library/jest-dom/extend-expect';
 import { configure as configureRTL } from '@testing-library/react';
-import * as Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-import { configure } from 'enzyme';
 import { noop } from 'lodash';
 
-const adapter = Adapter as any;
-
-configure({ adapter: new adapter.default() });
 configureRTL({ testIdAttribute: 'data-test' });
 
 // https://github.com/facebook/jest/issues/10784
@@ -41,6 +37,25 @@ Object.defineProperty(
 
 (global as any).__webpack_public_path__ = undefined;
 
+const originalConsoleError = console.error;
+const originalConsoleWarn = console.warn;
+
 beforeAll(() => {
     expect.hasAssertions();
+
+    console.error = (...args: unknown[]) => {
+        if (args.map(String).join().includes('Formik')) {
+            return;
+        }
+
+        originalConsoleError(...args);
+    };
+
+    console.warn = (...args: unknown[]) => {
+        if (args.map(String).join().includes('Formik')) {
+            return;
+        }
+
+        originalConsoleWarn(...args);
+    };
 });